changed the login components and simplified some of the code, thanks wilson
