+++ client/src/App.js
@@ -5,17 +5,24 @@ import MedicalProfile from "./component/pages/MedicalProfile"
 import NewPost from './component/pages/NewPost/NewPost'
 import PrintOut from './component/pages/PrintOut';
 import Navbar from "./component/Navbar";
+import CreateAccountPage from "./component/pages/CreateAccountPage";
 import Chart from './component/pages/Chart';
 function App() {
   return (
     <Fragment>
-       <Router>
-      <Navbar />
+       <Router>      
         <Switch>
+          <Route path="/">
+            <CreateAccountPage />
+          </Route>
+
           <Route absolute path="/medical-profile">
+          <Navbar />
             <MedicalProfile />
           </Route>
           <Route absolute path="/find-physician">
+          <Navbar />
+
             {/* <FindPhysician/> */}
           </Route>
           <Route absolute path="/chart">
